Refactor to get rid of master fragment
Move stuff from master fragment to main activity
